src/requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.0.0
numpy>=1.21.0
python-docx>=0.8.11
nltk>=3.8.0
scikit-learn>=1.2.0
tqdm>=4.64.0

# Метрики качества
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.0
scikit-learn>=1.2.0
torch>=2.0.0
transformers>=4.30.0

# Зависимости для API
Flask>=2.0.0
flask-cors>=3.0.0